requirements.txt
Pillow==3.1.1